Add service configs and dependencies
