+++ pyproject.toml
@@ -10,6 +10,7 @@ fastapi = "^0.62.0"
 uvicorn = "^0.13.1"
 ujson = "^4.0.1"
 requests = "^2.25.0"
+toolz = "^0.11.1"
 
 [tool.poetry.dev-dependencies]
 pep8 = "^1.7.1"
